refactor: "현재 시간표의 강의" 관련 기능만 CurrentTableLectureRepository 신설하여 추출
